+++ release-notes.md
@@ -2,7 +2,9 @@
 
 ## v1.1.1.dev0
 
-...
+Fixes for openssl loading (only relevant for historical signatures:
+    - Fix issue with libssl on MacOSX 10.15. by trying to load `libssl.35` first
+    - The `libeay32` library is only tried to be loaded on Windows platform
 
 ## v1.1.0